add abstractmethod flags
