+++ tests/webscraping/test_w3c.py
@@ -28,6 +28,7 @@ class TestW3CMessageParser:
         msg_parser = W3CMessageParser(
             website=True,
         )
+        print("------->", msg_parser.session)
         msg = msg_parser.from_url(
             list_name="public-test2",
             url=url_message,
